api/requirements.txt
amqp==5.0.2
asgiref==3.3.1
attrs==20.3.0
billiard==3.6.3.0
bleach==3.3.0
cachetools==4.1.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
cycler==0.10.0
defusedxml==0.7.0rc1
Django==3.1.4
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
django-user-sessions==1.7.1
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
docutils==0.16
drf-spectacular==0.14.0
ecdsa==0.14.1
environs==9.3.1
Flask==1.1.2
idna==2.10
importlib-metadata==3.7.2
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
jeepney==0.6.0
Jinja2==2.11.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jwt==1.2.0
keyring==23.0.0
kiwisolver==1.3.1
kombu==5.0.2
libsass==0.20.1
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.10.0
matplotlib==3.3.4
mpld3==0.5.2
msgpack==1.0.2
numpy==1.19.5
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
pandas==1.1.5
Pillow==8.1.2
pkginfo==1.7.0
presalytics>=0.6.2
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
py-gfm==1.0.2
pyasn1==0.4.8
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jose==3.2.0
python-json-logger==2.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
readme-renderer==29.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7.2
SecretStorage==3.3.1
semantic-version==2.8.5
signalrcore==0.8.8
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
tqdm==4.59.0
twine==3.3.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
wsgi-microservice-middleware==0.1.6
zipp==3.4.0
